Shiny App V1 obs/requirements.txt
# Shiny App – MBTA Red Line Tracker
# Install: pip install -r requirements.txt

shiny>=1.0.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
plotly>=5.18.0
shinywidgets>=0.3.0
polyline>=2.0.0
